Copied MockPlatform from Doctrine sources
The tests for this project had a dependency to mocks defined in
Doctrine test code. This is unfortunate, as this is not part of the
public api and can easily be changed, which it has.

Therefore has the class been copied to this code base. As the Doctrine
is currently licensed under the MIT license, the direct copying of that
code should not be a problem and in violation of any licensing issues.